requirements.txt
opencv-python>=4.9.0
ultralytics>=8.1.0
streamlit>=1.35.0
python-dotenv>=1.0.0
Pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
openai>=1.30.0
google-generativeai>=0.5.0
groq>=0.9.0
